refactor: refactored FFtFrame impl per platform
