+++ packages/react-native-audio-api/common/cpp/utils/FFTFrame.cpp
@@ -1,8 +1,15 @@
-#if defined(HAVE_ACCELERATE)
 #include "FFTFrame.h"
-#include "Accelerate/Accelerate.h"
+
+#if defined(HAVE_ACCELERATE)
+#include <Accelerate/Accelerate.h>
+#endif
+
+#if defined(ANDROID)
+#include <fftw3.h>
+#endif
 
 namespace audioapi {
+#if defined(HAVE_ACCELERATE)
 
 void FFTFrame::inverse(float *timeDomainData) {
   FFTSetup fftSetup_ = vDSP_create_fftsetup(log2Size_, FFT_RADIX2);
@@ -25,5 +32,25 @@ void FFTFrame::inverse(float *timeDomainData) {
 
   vDSP_destroy_fftsetup(fftSetup_);
 }
-} // namespace audioapi
+
+#elif defined(ANDROID)
+
+void FFTFrame::inverse(float *timeDomainData) {
+  fftwf_complex *freqDomainData = fftwf_alloc_complex(size_ / 2);
+  for (int i = 0; i < size_ / 2; i++) {
+    freqDomainData[i][0] = realData_[i];
+    freqDomainData[i][1] = imaginaryData_[i];
+  }
+
+  auto plan = fftwf_plan_dft_c2r_1d(
+      size_, freqDomainData, timeDomainData, FFTW_ESTIMATE);
+  fftwf_execute(plan);
+  fftwf_destroy_plan(plan);
+  fftwf_free(freqDomainData);
+
+  VectorMath::multiplyByScalar(
+      timeDomainData, 1.0f / static_cast<float>(size_), timeDomainData, size_);
+}
+
 #endif
+} // namespace audioapi
